Move color space transform methods into own class
